New implementation with state machine
