+++ platformio.ini
@@ -13,6 +13,8 @@ platform = espressif8266
 board = d1_mini_pro
 framework = arduino
 build_type = debug
+build_flags = -Isrc/lib
+monitor_filters = esp8266_exception_decoder
 
 board_build.ldscript = eagle.flash.8m7m.ld
 board_build.filesystem = littlefs
