--- src/MediaPlayer.h
@@ -1,6 +1,0 @@
-#ifndef SRC_MediaPlayer_H_
-#define SRC_MediaPlayer_H_
-
-#include "MediaPlayer.hpp"
-
-#endif  // SRC_MediaPlayer_H_
